+++ src/main/java/com/arakelian/retry/StopStrategy.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package com.github.rholder.retry;
+package com.arakelian.retry;
 
 /**
  * A strategy used to decide if a retryer must stop retrying after a failed attempt or not.
@@ -26,7 +26,8 @@ public interface StopStrategy {
     /**
      * Returns <code>true</code> if the retryer should stop retrying.
      *
-     * @param failedAttempt the previous failed {@code Attempt}
+     * @param failedAttempt
+     *            the previous failed {@code Attempt}
      * @return <code>true</code> if the retryer must stop, <code>false</code> otherwise
      */
     boolean shouldStop(Attempt failedAttempt);