+++ src/test/java/com/arakelian/retry/StopStrategiesTest.java
@@ -14,8 +14,9 @@
  * limitations under the License.
  */
 
-package com.github.rholder.retry;
+package com.arakelian.retry;
 
+import static java.util.concurrent.TimeUnit.MILLISECONDS;
 import static org.junit.Assert.assertFalse;
 import static org.junit.Assert.assertTrue;
 
@@ -24,6 +25,10 @@ import java.util.concurrent.TimeUnit;
 import org.junit.Test;
 
 public class StopStrategiesTest {
+
+    public Attempt<Boolean> failedAttempt(final long attemptNumber, final long delaySinceFirstAttempt) {
+        return new Retryer.ExceptionAttempt<>(new RuntimeException(), attemptNumber, delaySinceFirstAttempt);
+    }
 
     @Test
     public void testNeverStop() {
@@ -39,9 +44,9 @@ public class StopStrategiesTest {
 
     @Test
     public void testStopAfterDelayWithMilliseconds() {
-        assertFalse(StopStrategies.stopAfterDelay(1000L).shouldStop(failedAttempt(2, 999L)));
-        assertTrue(StopStrategies.stopAfterDelay(1000L).shouldStop(failedAttempt(2, 1000L)));
-        assertTrue(StopStrategies.stopAfterDelay(1000L).shouldStop(failedAttempt(2, 1001L)));
+        assertFalse(StopStrategies.stopAfterDelay(1000L, MILLISECONDS).shouldStop(failedAttempt(2, 999L)));
+        assertTrue(StopStrategies.stopAfterDelay(1000L, MILLISECONDS).shouldStop(failedAttempt(2, 1000L)));
+        assertTrue(StopStrategies.stopAfterDelay(1000L, MILLISECONDS).shouldStop(failedAttempt(2, 1001L)));
     }
 
     @Test
@@ -49,9 +54,5 @@ public class StopStrategiesTest {
         assertFalse(StopStrategies.stopAfterDelay(1, TimeUnit.SECONDS).shouldStop(failedAttempt(2, 999L)));
         assertTrue(StopStrategies.stopAfterDelay(1, TimeUnit.SECONDS).shouldStop(failedAttempt(2, 1000L)));
         assertTrue(StopStrategies.stopAfterDelay(1, TimeUnit.SECONDS).shouldStop(failedAttempt(2, 1001L)));
-    }
-
-    public Attempt<Boolean> failedAttempt(long attemptNumber, long delaySinceFirstAttempt) {
-        return new Retryer.ExceptionAttempt<Boolean>(new RuntimeException(), attemptNumber, delaySinceFirstAttempt);
     }
 }